Refactor nonblocking calibration code
