+++ src/sensors/nonblockingcalibration/NonBlockingCalibration.h
@@ -1,0 +1,317 @@
+/*
+	SlimeVR Code is placed under the MIT license
+	Copyright (c) 2024 Gorbit99 & SlimeVR Contributors
+
+	Permission is hereby granted, free of charge, to any person obtaining a copy
+	of this software and associated documentation files (the "Software"), to deal
+	in the Software without restriction, including without limitation the rights
+	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+	copies of the Software, and to permit persons to whom the Software is
+	furnished to do so, subject to the following conditions:
+
+	The above copyright notice and this permission notice shall be included in
+	all copies or substantial portions of the Software.
+
+	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+	THE SOFTWARE.
+*/
+
+#pragma once
+
+#include <vector3.h>
+
+#include <cstdint>
+#include <functional>
+#include <optional>
+
+#include "../../GlobalVars.h"
+#include "../../configuration/Configuration.h"
+#include "../../utils.h"
+#include "../SensorFusionRestDetect.h"
+#include "AccelBiasCalibrationStep.h"
+#include "GyroBiasCalibrationStep.h"
+#include "MotionlessCalibrationStep.h"
+#include "NullCalibrationStep.h"
+#include "SampleRateCalibrationStep.h"
+
+namespace SlimeVR::Sensors::NonBlockingCalibration {
+
+template <typename IMU>
+class NonBlockingCalibrator {
+public:
+	mutable SlimeVR::Logging::Logger logger{"DynamicCalibration"};
+
+	NonBlockingCalibrator(
+		SensorFusionRestDetect& fusion,
+		float accelScale,
+		IMU& imu,
+		uint8_t sensorId
+	)
+		: fusion{fusion}
+		, accelScale{accelScale}
+		, imu{imu}
+		, sensorId{sensorId} {}
+
+	void setup(Configuration::NonBlockingCalibrationConfig& baseCalibrationConfig) {
+		calibrationConfig = baseCalibrationConfig;
+		startupMillis = millis();
+
+		gyroBiasCalibrationStep.swapCalibrationIfNecessary();
+
+		computeNextCalibrationStep();
+
+		printCalibration();
+	}
+
+	void tick() {
+		if (skippedAStep && !lastTickRest && fusion.getRestDetected()) {
+			computeNextCalibrationStep();
+			skippedAStep = false;
+		}
+
+		if (millis() - startupMillis < initialStartupDelaySeconds * 1e3) {
+			return;
+		}
+
+		if (!fusion.getRestDetected() && currentStep->requiresRest()) {
+			if (isCalibrating) {
+				currentStep->cancel();
+				isCalibrating = false;
+			}
+
+			lastTickRest = fusion.getRestDetected();
+			return;
+		}
+
+		auto result = currentStep->tick();
+
+		switch (result) {
+			case CalibrationStep::TickResult::DONE:
+				stepCalibrationForward();
+				break;
+			case CalibrationStep::TickResult::SKIP:
+				stepCalibrationForward(false);
+				break;
+			case CalibrationStep::TickResult::CONTINUE:
+				break;
+		}
+
+		lastTickRest = fusion.getRestDetected();
+	}
+
+	void provideAccelSample(const int16_t accelSample[3]) {
+		currentStep->processAccelSample(accelSample);
+	}
+
+	void provideGyroSample(const int16_t gyroSample[3]) {
+		currentStep->processGyroSample(gyroSample);
+	}
+
+	void provideTempSample(float tempSample) {
+		currentStep->processTempSample(tempSample);
+	}
+
+private:
+	enum class CalibrationStepEnum {
+		NONE,
+		SAMPLING_RATE,
+		MOTIONLESS,
+		GYRO_BIAS,
+		ACCEL_BIAS,
+	};
+
+	void computeNextCalibrationStep() {
+		if (!calibrationConfig.sensorTimestepsCalibrated) {
+			nextCalibrationStep = CalibrationStepEnum::SAMPLING_RATE;
+			currentStep = &sampleRateCalibrationStep;
+		} else if (!calibrationConfig.motionlessCalibrated && HasMotionlessCalib) {
+			nextCalibrationStep = CalibrationStepEnum::MOTIONLESS;
+			currentStep = &motionlessCalibrationStep;
+		} else if (calibrationConfig.gyroPointsCalibrated == 0) {
+			nextCalibrationStep = CalibrationStepEnum::GYRO_BIAS;
+			currentStep = &gyroBiasCalibrationStep;
+		} else if (!accelBiasCalibrationStep.allAxesCalibrated()) {
+			nextCalibrationStep = CalibrationStepEnum::ACCEL_BIAS;
+			currentStep = &accelBiasCalibrationStep;
+		} else {
+			nextCalibrationStep = CalibrationStepEnum::GYRO_BIAS;
+			currentStep = &gyroBiasCalibrationStep;
+		}
+
+		currentStep->start();
+	}
+
+	void stepCalibrationForward(bool save = true) {
+		currentStep->cancel();
+		switch (nextCalibrationStep) {
+			case CalibrationStepEnum::NONE:
+				return;
+			case CalibrationStepEnum::SAMPLING_RATE:
+				nextCalibrationStep = CalibrationStepEnum::MOTIONLESS;
+				currentStep = &motionlessCalibrationStep;
+				if (save) {
+					printCalibration(CalibrationPrintFlags::TIMESTEPS);
+				}
+				break;
+			case CalibrationStepEnum::MOTIONLESS:
+				nextCalibrationStep = CalibrationStepEnum::GYRO_BIAS;
+				currentStep = &gyroBiasCalibrationStep;
+				if (save) {
+					printCalibration(CalibrationPrintFlags::MOTIONLESS);
+				}
+				break;
+			case CalibrationStepEnum::GYRO_BIAS:
+				if (calibrationConfig.gyroPointsCalibrated == 1) {
+					nextCalibrationStep = CalibrationStepEnum::ACCEL_BIAS;
+					currentStep = &accelBiasCalibrationStep;
+				}
+
+				if (save) {
+					printCalibration(CalibrationPrintFlags::GYRO_BIAS);
+				}
+				break;
+			case CalibrationStepEnum::ACCEL_BIAS:
+				nextCalibrationStep = CalibrationStepEnum::GYRO_BIAS;
+				currentStep = &gyroBiasCalibrationStep;
+
+				if (save) {
+					printCalibration(CalibrationPrintFlags::ACCEL_BIAS);
+				}
+
+				if (!accelBiasCalibrationStep.allAxesCalibrated()) {
+					skippedAStep = true;
+				}
+				break;
+		}
+
+		currentStep->start();
+
+		isCalibrating = false;
+
+		if (save) {
+			saveCalibration();
+		}
+	}
+
+	void saveCalibration() {
+		SlimeVR::Configuration::CalibrationConfig calibration;
+		calibration.type = SlimeVR::Configuration::CalibrationConfigType::NONBLOCKING;
+		calibration.data.nonblocking = calibrationConfig;
+		configuration.setCalibration(sensorId, calibration);
+		configuration.save();
+
+		ledManager.blink(100);
+	}
+
+	enum class CalibrationPrintFlags {
+		TIMESTEPS = 1,
+		MOTIONLESS = 2,
+		GYRO_BIAS = 4,
+		ACCEL_BIAS = 8,
+	};
+
+	static constexpr CalibrationPrintFlags PrintAllCalibration
+		= CalibrationPrintFlags::TIMESTEPS | CalibrationPrintFlags::MOTIONLESS
+		| CalibrationPrintFlags::GYRO_BIAS | CalibrationPrintFlags::ACCEL_BIAS;
+
+	void printCalibration(CalibrationPrintFlags toPrint = PrintAllCalibration) {
+		if (any(toPrint & CalibrationPrintFlags::TIMESTEPS)) {
+			if (calibrationConfig.sensorTimestepsCalibrated) {
+				logger.info(
+					"Calibrated timesteps: Accel %f, Gyro %f, Temperature %f",
+					calibrationConfig.A_Ts,
+					calibrationConfig.G_Ts,
+					calibrationConfig.T_Ts
+				);
+			} else {
+				logger.info("Sensor timesteps not calibrated");
+			}
+		}
+
+		if (HasMotionlessCalib && any(toPrint & CalibrationPrintFlags::MOTIONLESS)) {
+			if (calibrationConfig.motionlessCalibrated) {
+				logger.info("Motionless calibration done");
+			} else {
+				logger.info("Motionless calibration not done");
+			}
+		}
+
+		if (any(toPrint & CalibrationPrintFlags::GYRO_BIAS)) {
+			if (calibrationConfig.gyroPointsCalibrated != 0) {
+				logger.info(
+					"Calibrated gyro bias at %fC: %f %f %f",
+					calibrationConfig.gyroMeasurementTemperature1,
+					calibrationConfig.G_off1[0],
+					calibrationConfig.G_off1[1],
+					calibrationConfig.G_off1[2]
+				);
+			} else {
+				logger.info("Gyro bias not calibrated");
+			}
+
+			if (calibrationConfig.gyroPointsCalibrated == 2) {
+				logger.info(
+					"Calibrated gyro bias at %fC: %f %f %f",
+					calibrationConfig.gyroMeasurementTemperature2,
+					calibrationConfig.G_off2[0],
+					calibrationConfig.G_off2[1],
+					calibrationConfig.G_off2[2]
+				);
+			}
+		}
+
+		if (any(toPrint & CalibrationPrintFlags::ACCEL_BIAS)) {
+			if (accelBiasCalibrationStep.allAxesCalibrated()) {
+				logger.info(
+					"Calibrated accel bias: %f %f %f",
+					calibrationConfig.A_off[0],
+					calibrationConfig.A_off[1],
+					calibrationConfig.A_off[2]
+				);
+			} else if (accelBiasCalibrationStep.anyAxesCalibrated()) {
+				logger.info(
+					"Partially calibrated accel bias: %f %f %f",
+					calibrationConfig.A_off[0],
+					calibrationConfig.A_off[1],
+					calibrationConfig.A_off[2]
+				);
+			} else {
+				logger.info("Accel bias not calibrated");
+			}
+		}
+	}
+
+	static constexpr bool HasMotionlessCalib
+		= requires(IMU& i) { typename IMU::MotionlessCalibrationData; };
+
+	SensorFusionRestDetect& fusion;
+	IMU& imu;
+	float accelScale;
+	uint8_t sensorId;
+
+	CalibrationStepEnum nextCalibrationStep = CalibrationStepEnum::MOTIONLESS;
+
+	static constexpr float initialStartupDelaySeconds = 5;
+	uint64_t startupMillis;
+
+	SlimeVR::Configuration::NonBlockingCalibrationConfig calibrationConfig;
+
+	SampleRateCalibrationStep sampleRateCalibrationStep{calibrationConfig};
+	MotionlessCalibrationStep<IMU> motionlessCalibrationStep{calibrationConfig, imu};
+	GyroBiasCalibrationStep gyroBiasCalibrationStep{calibrationConfig};
+	AccelBiasCalibrationStep accelBiasCalibrationStep{calibrationConfig, accelScale};
+	NullCalibrationStep nullCalibrationStep{calibrationConfig};
+
+	CalibrationStep* currentStep = &nullCalibrationStep;
+
+	bool isCalibrating = false;
+	bool skippedAStep = false;
+	bool lastTickRest = false;
+};
+
+}  // namespace SlimeVR::Sensors::NonBlockingCalibration
